Updated abstract method validate to be boolean
Overrode method in StandScoutAutonFragment and tracked value so that all
errors get thrown back to user, rather than method terminating as soon
as it hits its return call. Please tell me if this is correct, because
I will be using the same template for the teleop and postgame sections of
the DataCollection.